Value parser to float

+++ src/pages/follow-up-analysis/follow-up-analysis-table/FollowUpAnalysisTable.js
@@ -44,7 +44,7 @@ class FollowUpAnalysisTable extends Component {
         period: e.period.name,
         min: e.min,
         max: e.max,
-        value: Number.parseInt(e.value, 10),
+        value: Number.parseFloat(e.value, 10),
         marked: !e.followup,
         comment: e.comment,
     });
